requirements.txt
requests
google-generativeai
newsapi-python
python-dotenv